PIM-3832: move new 1.4 classes from BaseConnectorBundle to ConnectorBundle
