+++ src/Pim/Bundle/ConnectorBundle/DependencyInjection/PimConnectorExtension.php
@@ -22,7 +22,9 @@ class PimConnectorExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader->load('models.yml');
         $loader->load('processors.yml');
+        $loader->load('repositories.yml');
         $loader->load('writers.yml');
     }
 }
