+++ src/Pim/Bundle/ConnectorBundle/Resources/config/models.yml
@@ -1,0 +1,2 @@
+parameters:
+    pim_base_connector.model.job_configuration.class: Pim\Component\Connector\Model\JobConfiguration
